reworked message formatting to deal with warnings from Emscripten compiler

+++ OOps/diskin2.c
@@ -467,10 +467,10 @@ static int diskin2_init_(CSOUND *csound, DISKIN2 *p, int stringname)
       if (UNLIKELY((csound->oparms_.msglevel & 7) == 7)) {
         csound->Message(csound, Str("diskin2: opened (asynchronously) '%s':\n"
                                     "         %d Hz, %d channel(s), "
-                                    "%ld sample frames\n"),
+                                    "%d sample frames\n"),
                         csound->GetFileName(fd),
-                        (int)sfinfo.samplerate, (int)sfinfo.channels,
-                        (int32) sfinfo.frames);
+                        sfinfo.samplerate, sfinfo.channels,
+                        sfinfo.frames);
       }
     }
     else {
@@ -481,10 +481,10 @@ static int diskin2_init_(CSOUND *csound, DISKIN2 *p, int stringname)
       if (UNLIKELY((csound->oparms_.msglevel & 7) == 7)) {
         csound->Message(csound, Str("diskin2: opened '%s':\n"
                                     "         %d Hz, %d channel(s), "
-                                    "%ld sample frames\n"),
+                                    "%d sample frames\n"),
                         csound->GetFileName(fd),
-                        (int)sfinfo.samplerate, (int)sfinfo.channels,
-                        (int32) sfinfo.frames);
+                        sfinfo.samplerate, sfinfo.channels,
+                        sfinfo.frames);
       }
     }
 
@@ -1741,10 +1741,10 @@ static int diskin2_init_array(CSOUND *csound, DISKIN2_ARRAY *p, int stringname)
       if (UNLIKELY((csound->oparms_.msglevel & 7) == 7)) {
         csound->Message(csound, Str("diskin2: opened (asynchronously) '%s':\n"
                                     "         %d Hz, %d channel(s), "
-                                    "%ld sample frames\n"),
+                                    "%d sample frames\n"),
                         csound->GetFileName(fd),
-                        (int)sfinfo.samplerate, (int)sfinfo.channels,
-                        (int32) sfinfo.frames);
+                        sfinfo.samplerate, sfinfo.channels,
+                        sfinfo.frames);
       }
     }
     else {
@@ -1755,10 +1755,10 @@ static int diskin2_init_array(CSOUND *csound, DISKIN2_ARRAY *p, int stringname)
       if (UNLIKELY((csound->oparms_.msglevel & 7) == 7)) {
         csound->Message(csound, Str("diskin2: opened '%s':\n"
                                     "         %d Hz, %d channel(s), "
-                                    "%ld sample frames\n"),
+                                    "%d sample frames\n"),
                         csound->GetFileName(fd),
-                        (int)sfinfo.samplerate, (int)sfinfo.channels,
-                        (int32) sfinfo.frames);
+                        sfinfo.samplerate, sfinfo.channels,
+                        sfinfo.frames);
       }
     }
